Bug 544147: [Release Wizard] Add buttons to select/deselect all packages
In the case of multiple packages, add buttons to select all and
deselect all packages for update.  Also, on selection of a package
to update, pre-fill a suggested version number if the label to match
looks like a possibly dot-segment version number, possibly following
an underscore ('_'), hyphen ('-'), or a 'v' that doesn't end a word.

Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>
Change-Id: I6321c866e6f0a2388fe84e4698b33450c5f03859
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>

Former-commit-id: fb9781e762a75e17d7935ad937704c042ee70cdf